requirements.txt
flask
google-generativeai
pandas 
python-dotenv 
psycopg2